+++ src/client/providers/quickFixes/fixLaunchJson.ts
