+++ src/client/providers/serviceRegistry.ts
@@ -3,10 +3,13 @@
 
 'use strict';
 
+import { IExtensionSingleActivationService } from '../activation/types';
 import { IServiceManager } from '../ioc/types';
 import { SortImportsEditingProvider } from './importSortProvider';
+import { QuickFixService } from './quickFixes/main';
 import { ISortImportsEditingProvider } from './types';
 
 export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<ISortImportsEditingProvider>(ISortImportsEditingProvider, SortImportsEditingProvider);
+    serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, QuickFixService);
 }
