+++ src/test/providers/quickFixes/main.unit.test.ts
@@ -7,8 +7,8 @@
 import { assert, expect } from 'chai';
 import rewiremock from 'rewiremock';
 import { CodeActionProvider, CodeActionProviderMetadata, DocumentSelector } from 'vscode';
-import { QuickFixLaunchJson } from '../../../client/language/quickFixes/fixLaunchJson';
-import { QuickFixService } from '../../../client/language/quickFixes/main';
+import { QuickFixLaunchJson } from '../../../client/providers/quickFixes/fixLaunchJson';
+import { QuickFixService } from '../../../client/providers/quickFixes/main';
 
 suite('Quick fix service', async () => {
     setup(() => {
