+++ src/test/providers/serviceRegistry.unit.test.ts
@@ -4,9 +4,11 @@
 'use strict';
 
 import { instance, mock, verify } from 'ts-mockito';
+import { IExtensionSingleActivationService } from '../../client/activation/types';
 import { ServiceManager } from '../../client/ioc/serviceManager';
 import { IServiceManager } from '../../client/ioc/types';
 import { SortImportsEditingProvider } from '../../client/providers/importSortProvider';
+import { QuickFixService } from '../../client/providers/quickFixes/main';
 import { registerTypes } from '../../client/providers/serviceRegistry';
 import { ISortImportsEditingProvider } from '../../client/providers/types';
 
@@ -23,6 +25,12 @@ suite('Common Providers Service Registry', () => {
             serviceManager.addSingleton<ISortImportsEditingProvider>(
                 ISortImportsEditingProvider,
                 SortImportsEditingProvider
+            )
+        ).once();
+        verify(
+            serviceManager.addSingleton<IExtensionSingleActivationService>(
+                IExtensionSingleActivationService,
+                QuickFixService
             )
         ).once();
     });
